% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_advanced_search.r
\name{am_advanced_search}
\alias{am_advanced_search}
\title{Get Advanced Search}
\usage{
am_advanced_search(
  release_date_gte,
  release_date_lte,
  country,
  store,
  size,
  screenshots = NULL,
  search = NULL,
  description = NULL,
  sort,
  sort_asc = NULL,
  category_ids = NULL,
  tag_id = NULL,
  publisher_place = NULL,
  signature,
  accept = "text/csv"
)
}
\arguments{
\item{release_date_gte}{Release Date Gte (minimum date is 2015-01-01)}

\item{release_date_lte}{Release Date Lte (minimum date is 2015-01-01)}

\item{country}{Alpha-2 country code}

\item{store}{Store code}

\item{size}{The number of returned items}

\item{screenshots}{If to include screenshots}

\item{search}{Search query}

\item{description}{Description search query}

\item{sort}{Type of a TOP}

\item{sort_asc}{Ascending or descending sorting}

\item{category_ids}{Category IDs}

\item{tag_id}{Tag IDs}

\item{publisher_place}{Publisher place}

\item{signature}{Authentication token}

\item{accept}{Accept header value}
}
\description{
Get Advanced Search <https://api.appmagic.rocks/v1/docs#operation/advancedSearch>
}
