% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_app.r
\name{am_app}
\alias{am_app}
\title{Get Application by Store Application ID}
\usage{
am_app(store, store_application_id, signature, accept = "text/csv")
}
\arguments{
\item{store}{Store code}

\item{store_application_id}{Store application ID}

\item{signature}{Authentication token}

\item{accept}{Accept header value}
}
\description{
Get Application by Store Application ID <https://api.appmagic.rocks/v1/docs#operation/getApplication>
}
