% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_app_info.r
\name{am_app_info}
\alias{am_app_info}
\title{Get Application Info by Store Application ID}
\usage{
am_app_info(store, store_application_id, signature)
}
\arguments{
\item{store}{Store code}

\item{store_application_id}{Store application ID}

\item{signature}{Authentication token}
}
\description{
Get Application Info by Store Application ID <https://api.appmagic.rocks/v1/docs#operation/getApplicationInfo>
}
