% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_app_screenshots.r
\name{am_app_screenshots}
\alias{am_app_screenshots}
\title{Get Application Screenshots by Store Application ID}
\usage{
am_app_screenshots(store, store_application_id, signature, accept = "text/csv")
}
\arguments{
\item{store}{Store code}

\item{store_application_id}{Store application ID}

\item{signature}{Authentication token}

\item{accept}{Accept header value}
}
\description{
Get Application Screenshots by Store Application ID <https://api.appmagic.rocks/v1/docs#operation/getApplicationScreenshots>
}
