% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_apps.r
\name{am_apps}
\alias{am_apps}
\title{Get Applications by Search Query}
\usage{
am_apps(search, limit = NULL, offset = NULL, signature)
}
\arguments{
\item{search}{Search query}

\item{limit}{The number of returned items}

\item{offset}{Offset of returned items}

\item{signature}{Authentication token}
}
\description{
Get Applications by Search Query <https://api.appmagic.rocks/v1/docs#operation/getApplications>
}
