% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_apps_ads.r
\name{am_apps_ads}
\alias{am_apps_ads}
\title{Get Application Ads}
\usage{
am_apps_ads(
  appIds,
  country,
  dateFrom,
  dateTo,
  sort,
  aggregation,
  count,
  offset,
  adNetwork,
  platform,
  type,
  signature
)
}
\arguments{
\item{appIds}{Application store ID}

\item{country}{Country}

\item{dateFrom}{Start date (minimum date is 2015-01-01)}

\item{dateTo}{End date (minimum date is 2015-01-01)}

\item{sort}{Sorting type}

\item{aggregation}{Aggregation type}

\item{count}{The number of returned items}

\item{offset}{Offset of returned items}

\item{adNetwork}{Ad network}

\item{platform}{Platform}

\item{type}{Ad type}

\item{signature}{Authentication token}
}
\description{
Get Application Ads
}
