% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_apps_ids.r
\name{am_apps_ids}
\alias{am_apps_ids}
\title{Get Applications Info by Store Application IDs}
\usage{
am_apps_ids(ids_array, signature, accept = "text/csv")
}
\arguments{
\item{ids_array}{List of lists with store IDs and store applications IDs}

\item{signature}{Authentication token}

\item{accept}{Accept header value}
}
\description{
Get Applications Info by Store Application IDs <https://api.appmagic.rocks/v1/docs#operation/getApplicationsByIds>
}
