% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_auth.r
\name{am_auth}
\alias{am_auth}
\title{Get AppMagic API Authentication token}
\usage{
am_auth(name, password)
}
\arguments{
\item{name}{Account name}

\item{password}{Account password}
}
\description{
Get AppMagic API Authentication token <https://api.appmagic.rocks/v1/docs#section/Authentication>
}
