% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_contacts_companies.r
\name{am_contacts_companies}
\alias{am_contacts_companies}
\title{Get LinkedIn Companies from United Publisher ID}
\usage{
am_contacts_companies(id, signature)
}
\arguments{
\item{id}{United publisher ID}

\item{signature}{Authentication token}
}
\description{
Get LinkedIn Companies from United Publisher ID <https://api.appmagic.rocks/v1/docs#operation/getCompanies>
}
