% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_contacts_profiles.r
\name{am_contacts_profiles}
\alias{am_contacts_profiles}
\title{Get Profiles from Company ID}
\usage{
am_contacts_profiles(
  company_id,
  page = NULL,
  person_not_titles = NULL,
  person_titles = NULL,
  signature
)
}
\arguments{
\item{company_id}{Company ID}

\item{page}{Page number}

\item{person_not_titles}{Titles to exclude}

\item{person_titles}{Titles to include}

\item{signature}{Authentication token}
}
\description{
Get Profiles from Company ID <https://api.appmagic.rocks/v1/docs#operation/getProfiles>
}
