% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_history_apps.r
\name{am_history_apps}
\alias{am_history_apps}
\title{Get Data for Application}
\usage{
am_history_apps(
  aggregation = NULL,
  date_from = NULL,
  date_to = NULL,
  store,
  store_application_id,
  country,
  category_id = NULL,
  signature
)
}
\arguments{
\item{aggregation}{Period of data summarization}

\item{date_from}{Start date for data (minimum date is 2015-01-01)}

\item{date_to}{End date for data (minimum date is 2015-01-01)}

\item{store}{Store code}

\item{store_application_id}{Store app ID}

\item{country}{Alpha-2 country code}

\item{category_id}{Category ID}

\item{signature}{Authentication token}
}
\description{
Get Data for Application <https://api.appmagic.rocks/v1/docs#operation/getApplicationHistory>
}
