% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_history_united_app.r
\name{am_history_united_app}
\alias{am_history_united_app}
\title{Get Data for United Application}
\usage{
am_history_united_app(
  aggregation = NULL,
  date_from = NULL,
  date_to = NULL,
  store,
  united_application_id,
  country,
  signature
)
}
\arguments{
\item{aggregation}{Period of data summarization}

\item{date_from}{Start date for data (minimum date is 2015-01-01)}

\item{date_to}{End date for data (minimum date is 2015-01-01)}

\item{store}{Store code}

\item{united_application_id}{United app ID}

\item{country}{Alpha-2 country code}

\item{signature}{Authentication token}
}
\description{
Get Data for United Application <https://api.appmagic.rocks/v1/docs#operation/getUnitedApplicationHistory>
}
