% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_publisher.r
\name{am_publisher}
\alias{am_publisher}
\title{Get Publisher by Store Publisher ID}
\usage{
am_publisher(store, store_publisher_id, signature)
}
\arguments{
\item{store}{Store code}

\item{store_publisher_id}{Store publisher ID}

\item{signature}{Authentication token}
}
\description{
Get Publisher by Store Publisher ID <https://api.appmagic.rocks/v1/docs#operation/getPublisher>
}
