% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_publishers.r
\name{am_publishers}
\alias{am_publishers}
\title{Get Publishers by Search Query}
\usage{
am_publishers(search, limit = NULL, offset = NULL, signature)
}
\arguments{
\item{search}{Search query}

\item{limit}{The number of returned items}

\item{offset}{Offset of returned items}

\item{signature}{Authentication token}
}
\description{
Get Publishers by Search Query <https://api.appmagic.rocks/v1/docs#operation/getPublishers>
}
