% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_tags.r
\name{am_tags}
\alias{am_tags}
\title{Get AppMagic Tags}
\usage{
am_tags(signature)
}
\arguments{
\item{signature}{Authentication token}
}
\description{
Get AppMagic Tags <https://api.appmagic.rocks/v1/docs#operation/getTags>
}
