% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_top_apps.r
\name{am_top_apps}
\alias{am_top_apps}
\title{Get Top Applications}
\usage{
am_top_apps(
  aggregation = NULL,
  sort,
  store,
  country,
  category_id = NULL,
  date,
  limit = NULL,
  tag_id = NULL,
  signature
)
}
\arguments{
\item{aggregation}{Period of data summarization}

\item{sort}{Type of TOP}

\item{store}{Store code}

\item{country}{Alpha-2 country code}

\item{category_id}{Category ID}

\item{date}{Date of TOP (minimum date is 2015-01-01)}

\item{limit}{The number of returned items}

\item{tag_id}{Tag ID}

\item{signature}{Authentication token}
}
\description{
Get Top Applications
}
