% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_top_ltv.r
\name{am_top_ltv}
\alias{am_top_ltv}
\title{Get Top LTV}
\usage{
am_top_ltv(tag_id = NULL, store, limit = NULL, country, signature)
}
\arguments{
\item{tag_id}{Tag ID}

\item{store}{Store code}

\item{limit}{The number of returned items}

\item{country}{Alpha-2 country code}

\item{signature}{Authentication token}
}
\description{
Get Top LTV
}
