% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_top_publishers.r
\name{am_top_publishers}
\alias{am_top_publishers}
\title{Get Top Publishers}
\usage{
am_top_publishers(
  date,
  store,
  aggregation,
  tag_id = NULL,
  headquarters = NULL,
  countries,
  limit = NULL,
  sort,
  signature
)
}
\arguments{
\item{date}{Date of TOP (minimum date is 2015-01-01)}

\item{store}{Store code}

\item{aggregation}{Period of data summarization}

\item{tag_id}{Tag ID}

\item{headquarters}{Publishers headquarters country}

\item{countries}{Alpha-2 country code}

\item{limit}{The number of returned items}

\item{sort}{Sort by downloads, revenue or featuring}

\item{signature}{Authentication token}
}
\description{
Get Top Publishers <https://api.appmagic.rocks/v1/docs#operation/getTopPublishers>
}
