% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_top_soft_launches.r
\name{am_top_soft_launches}
\alias{am_top_soft_launches}
\title{Get Top Soft Launches}
\usage{
am_top_soft_launches(
  date_from,
  date_to,
  include_countries,
  exclude_countries = NULL,
  store,
  category_ids = NULL,
  only_global_publisher = NULL,
  top_publisher = NULL,
  sort,
  limit = NULL,
  signature
)
}
\arguments{
\item{date_from}{Start date of TOP (minimum date is 2015-01-01)}

\item{date_to}{End date of TOP (minimum date is 2015-01-01)}

\item{include_countries}{Include countries of soft launch}

\item{exclude_countries}{Exclude countries of soft launch}

\item{store}{Store code}

\item{category_ids}{Category IDs}

\item{only_global_publisher}{To include global publishers only}

\item{top_publisher}{Publisher rank}

\item{sort}{Sort by downloads or revenue}

\item{limit}{The number of returned items}

\item{signature}{Authentication token}
}
\description{
Get Top Soft Launches <https://api.appmagic.rocks/v1/docs#operation/getTopSoftLaunches>
}
