% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_top_trending.r
\name{am_top_trending}
\alias{am_top_trending}
\title{Get Top Trending Applications}
\usage{
am_top_trending(
  store,
  type,
  country,
  date_from,
  date_to,
  limit = NULL,
  depth = NULL,
  method = NULL,
  tag_id = NULL,
  category_ids = NULL,
  headquarter = NULL,
  new_only = NULL,
  kind = NULL,
  signature
)
}
\arguments{
\item{store}{Store code}

\item{type}{Type of a top (top free or top grossing)}

\item{country}{Alpha-2 country code}

\item{date_from}{Start date of TOP (minimum date is 2015-01-01)}

\item{date_to}{End date of TOP (minimum date is 2015-01-01)}

\item{limit}{The number of returned items}

\item{depth}{Top app analysis limit}

\item{method}{Type of a top (gainers or losers)}

\item{tag_id}{Tag IDs}

\item{category_ids}{Category IDs}

\item{headquarter}{Alpha-2 country code of application publisher}

\item{new_only}{Select apps released in specific time period or include others}

\item{kind}{Type of applications (all, games or apps)}

\item{signature}{Authentication token}
}
\description{
Get Top Trending Applications
}
