% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_top_united_apps.r
\name{am_top_united_apps}
\alias{am_top_united_apps}
\title{Get Top United Applications}
\usage{
am_top_united_apps(
  aggregation = NULL,
  sort,
  store,
  country,
  date,
  limit = NULL,
  tag_id = NULL,
  signature
)
}
\arguments{
\item{aggregation}{Period of data summarization}

\item{sort}{Type of TOP}

\item{store}{Store code}

\item{country}{Alpha-2 country code}

\item{date}{Date of TOP (minimum date is 2015-01-01)}

\item{limit}{The number of returned items}

\item{tag_id}{Tag ID}

\item{signature}{Authentication token}
}
\description{
Get Top United Applications <https://api.appmagic.rocks/v1/docs#operation/getTopUnitedApplications>
}
