% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_united_app.r
\name{am_united_app}
\alias{am_united_app}
\title{Get United Application by United Application ID}
\usage{
am_united_app(id, signature)
}
\arguments{
\item{id}{United application ID}

\item{signature}{Authentication token}
}
\description{
Get United Application by United Application ID <https://api.appmagic.rocks/v1/docs#operation/getUnitedApplication>
}
