% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_united_app_ltv.r
\name{am_united_app_ltv}
\alias{am_united_app_ltv}
\title{Get United Application LTV by United Application ID}
\usage{
am_united_app_ltv(id, signature)
}
\arguments{
\item{id}{United application ID}

\item{signature}{Authentication token}
}
\description{
Get United Application LTV by United Application ID <https://api.appmagic.rocks/v1/docs#operation/getUnitedApplicationLTV>
}
