% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_united_apps.r
\name{am_united_apps}
\alias{am_united_apps}
\title{Get United Applications by Search Query}
\usage{
am_united_apps(search, limit = NULL, offset = NULL, signature)
}
\arguments{
\item{search}{Search query}

\item{limit}{The number of returned items}

\item{offset}{Offset of returned items}

\item{signature}{Authentication token}
}
\description{
Get United Applications by Search Query <https://api.appmagic.rocks/v1/docs#operation/getUnitedApplications>
}
