% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_united_apps_app_ids.r
\name{am_united_apps_app_ids}
\alias{am_united_apps_app_ids}
\title{Get United Applications Info by Store Application IDs}
\usage{
am_united_apps_app_ids(ids_array, signature)
}
\arguments{
\item{ids_array}{List of lists with store IDs and store applications IDs}

\item{signature}{Authentication token}
}
\description{
Get United Applications Info by Store Application IDs <https://api.appmagic.rocks/v1/docs#operation/getUnitedApplicationsByAppIds>
}
