% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_united_apps_ids.r
\name{am_united_apps_ids}
\alias{am_united_apps_ids}
\title{Get United Applications Info by United Applications IDs}
\usage{
am_united_apps_ids(ids, signature, accept = "text/csv")
}
\arguments{
\item{ids}{Vector of united applications IDs}

\item{signature}{Authentication token}

\item{accept}{Accept header value}
}
\description{
Get United Applications Info by United Applications IDs <https://api.appmagic.rocks/v1/docs#operation/getUnitedApplicationsByIds>
}
