% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_united_publisher.r
\name{am_united_publisher}
\alias{am_united_publisher}
\title{Get Data for United Publisher}
\usage{
am_united_publisher(id, signature)
}
\arguments{
\item{id}{United publisher ID}

\item{signature}{Authentication token}
}
\description{
Get Data for United Publisher <https://api.appmagic.rocks/v1/docs#operation/getUnitedPublisher>
}
