% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_united_publisher_united_apps.r
\name{am_united_publisher_united_apps}
\alias{am_united_publisher_united_apps}
\title{Get United Applications for United Publisher}
\usage{
am_united_publisher_united_apps(id, signature)
}
\arguments{
\item{id}{United publisher ID}

\item{signature}{Authentication token}
}
\description{
Get United Applications for United Publisher <https://api.appmagic.rocks/v1/docs#operation/getUnitedPublisherUnitedApplications>
}
