% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_united_publishers.r
\name{am_united_publishers}
\alias{am_united_publishers}
\title{Search for United Publishers}
\usage{
am_united_publishers(search, signature)
}
\arguments{
\item{search}{Search query}

\item{signature}{Authentication token}
}
\description{
Search for United Publishers <https://api.appmagic.rocks/v1/docs#operation/getUnitedPublishers>
}
