% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_united_publishers_ids.r
\name{am_united_publishers_ids}
\alias{am_united_publishers_ids}
\title{Get United Publishers Info by United Publishers IDs}
\usage{
am_united_publishers_ids(ids, signature)
}
\arguments{
\item{ids}{Vector of united publishers IDs}

\item{signature}{Authentication token}
}
\description{
Get United Publishers Info by United Publishers IDs <https://api.appmagic.rocks/v1/docs#operation/getUnitedPublishersByIds>
}
