% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_app_similarity.r
\name{am_app_similarity}
\alias{am_app_similarity}
\title{Get Application Similarity by Store Application ID}
\usage{
am_app_similarity(store, store_application_id, signature, accept = "text/csv")
}
\arguments{
\item{store}{Store code}

\item{store_application_id}{Store application ID}

\item{signature}{Authentication token}

\item{accept}{Accept header value}
}
\description{
Get Application Similarity by Store Application ID <https://api.appmagic.rocks/v1/docs#operation/getApplicationSimilarity>
}
