% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_united_apps_ids.r
\name{am_united_apps_ids}
\alias{am_united_apps_ids}
\title{Get United Applications Info by United Applications IDs}
\usage{
am_united_apps_ids(ids, signature)
}
\arguments{
\item{ids}{Vector of united applications IDs}

\item{signature}{Authentication token}
}
\description{
Get United Applications Info by United Applications IDs
}
